Feat/review/11 (#13)
* add: 알림 추가

* add: 로그 추가

* 자료형

* 로깅 추가

+++ src/main/java/com/example/template/domain/college/dto/CollegeResponseDto.java
@@ -14,7 +14,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CollegeResponseDto {
-    private Long id;
+    private int id;
     private String name;
     private String headmaster;
     private String contactInfo;
@@ -23,7 +23,7 @@ public class CollegeResponseDto {
     private Boolean favorites;
 
     public CollegeResponseDto(College college) {
-        id = college.getId();
+        id = (int) college.getId();
         name = college.getName();
         headmaster = college.getHeadmaster();
         contactInfo = college.getContactNo();
